Improved the structure. Implemented an internal message sender.
Fixed some issues:

GCM Can send notifications and data:
Data Messages:

  {
    "to" : "bk3RNwTe3H0:CI2k_HHwgIpoDKCIZvvDMExUdFQ3P1...",
    "data" : {
      "body" : "great match!",
      "title" : "Portugal vs. Denmark",
      "icon" : "myicon"
    }
  }

  Notifications:
    {
      "to" : "bk3RNwTe3H0:CI2k_HHwgIpoDKCIZvvDMExUdFQ3P1...",
      "notification" : {
        "body" : "great match!",
        "title" : "Portugal vs. Denmark",
        "icon" : "myicon"
      }
    }

Data Messages are are better for sending customised notifications.
If we use the notification object in gcm sending - when the app is closed the system takes over the notifications and we issue un-customised notifications.
Thus for customised notifications always use data notifications.